css/css-flexbox/flexbox_fragmentation-propagation.html

<!DOCTYPE html>
<html lang="en">
<head>

  <title>Flexbox | Fragmentation propagation</title>
  <meta name="assert" content="Fragmentation propagation">
  <link rel="help" href="https://drafts.csswg.org/css-flexbox-1/#pagination">
  <link rel="author" title="Clayton Salem" href="claytonjsalem@gmail.com">
  <style>
    .item {
      background-color: brown;
      width: 96px;
      height: 96px;
      border: solid black 2px;
    }
    #container {
      display: flex;
      flex-direction: column;
      flex-wrap: wrap;
      height: 299px;
      padding: 5px;
      border: solid 7px #334488;
    }
    #container-break-before .item {
      break-before: always;
    }

  </style>
</head>
<body>
  <div id="container-break-before">
    <div class="item">1</div>
    <div class="item">2</div>
    <div class="item">3</div>
  </div>

  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script>
    // These tests are all assuming the left to right writing mode.
    test(function() {
      const container = document.querySelector("#container-break-before");
      const containerStyles = window.getComputedStyle(container);
      const containerRect = container.getBoundingClientRect();
      const firstItemRect = document.querySelector("#container-break-before .item:first-child").getBoundingClientRect();

      assert_equals(
        pxRemove(containerStyles.paddingTop) + pxRemove(containerStyles.borderTopWidth),
        diff(containerRect.top,firstItemRect.top)
      )
      assert_equals(
        pxRemove(containerStyles.paddingLeft) + pxRemove(containerStyles.borderLeftWidth),
        diff(containerRect.left,firstItemRect.left)
      )

    }, "The sum of the padding and border-width of the container are equal to \
    the difference between the top and left side of the conatiner and first rectangles.")

    function pxRemove(pxString) {
      return parseInt(pxString, 10);
    }
    function diff(a,b){
      return Math.abs(a-b);
    }
  </script>
</body>
</html>
